+++ src/paste_replaced/extension.cljs
@@ -12,10 +12,10 @@
     (.push (.-subscriptions context) disposable)))
 
 (defn- clear-disposables! []
-  (swap! db/!app-db assoc :disposables [])
-  (p/run! (fn [^js disposable]
-            (.dispose disposable))
-          (:disposables @db/!app-db)))
+  (-> (p/run! (fn [^js disposable]
+                (.dispose disposable))
+              (:disposables @db/!app-db))
+      (.then (fn [] (swap! db/!app-db assoc :disposables [])))))
 
 (def api (jsify {:getContextValue (fn [k]
                                     (when-contexts/context k))}))
@@ -26,16 +26,22 @@
     (swap! db/!app-db assoc
            :output-channel (vscode/window.createOutputChannel "Paste Replaced")
            :extension-context context
-           :workspace-root-path vscode/workspace.rootPath
+           :workspace-root-path (some-> (js->clj vscode/workspace.workspaceFolders :keywordize-keys true)
+                                        first)
            :typing-interrupted? false)
-    (utils/sayln "Paste Replace activating"))
-  (let [{:keys [extension-context]} @db/!app-db]
-    (register-command! extension-context "paste-replaced.paste" #'replacer/paste-replaced!+)
-    (register-command! extension-context "paste-replaced.pasteSelectionReplaced" #'replacer/select-and-paste-replaced!+)
-    (register-command! extension-context "paste-replaced.interruptTyping" #'replacer/interrupt-typing!)
-    (when-contexts/set-context!+ ::when-contexts/paste-replaced.isActive true)
-    (js/console.info "Paste Replace activate END")
-    api))
+    (utils/sayln "Paste Replace activating..."))
+  (try (let [{:keys [extension-context]} @db/!app-db]
+        (register-command! extension-context "paste-replaced.paste" #'replacer/paste-replaced!+)
+        (register-command! extension-context "paste-replaced.pasteSelectionReplaced" #'replacer/select-and-paste-replaced!+)
+        (register-command! extension-context "paste-replaced.interruptTyping" #'replacer/interrupt-typing!)
+        (when-contexts/set-context!+ ::when-contexts/paste-replaced.isActive true)
+        (js/console.info "Paste Replace activate END")
+        (utils/sayln "Paste Replace activation done"))
+      (catch :default e
+        (utils/say-error (str "Paste Replace activation failed: " (.-message e) ", see Development Console for stack trace"))
+        (throw e)))
+  
+  api)
 
 (defn ^:export deactivate []
   (when-contexts/set-context!+ ::when-contexts/paste-replaced.isActive false)
@@ -43,12 +49,14 @@
 
 
 (defn before [done]
+  (js/console.log "shadow-cljs before reloading paste-replaced")
   (-> (clear-disposables!)
       (p/then done)))
 
 (defn after []
   (info "shadow-cljs reloaded paste-replaced")
-  (js/console.log "shadow-cljs reloaded paste-replaced"))
+  (js/console.log "shadow-cljs reloaded paste-replaced")
+  (activate nil))
 
 (comment
   #_{:clj-kondo/ignore [:clojure-lsp/unused-public-var]}